refactor: remove UpsertBowl modal feature
