+++ app/bowls/new/page.tsx
@@ -4,7 +4,7 @@ import { useRouter } from "next/navigation";
 import { Button } from "@heroui/react";
 import { Icon } from "@iconify/react";
 
-import { BowlForm } from "@/features/upsert-bowl/ui/bowl-form";
+import { BowlForm } from "@/features/upsert-bowl";
 import { useBowls, type Bowl } from "@/entities/bowl";
 
 export type NewBowlPageProps = {};
